refactor: 가독성 및 컴포넌트 분리
