+++ src/common/global/SubFooter.tsx
@@ -1,39 +1,32 @@
-import React from 'react';
+import AppLinkBox from './AppLinkBox';
+import { InnerBox } from '../layouts';
 import styled from 'styled-components';
-import AppLinkBox from '../AppLinkBox';
+
+const SubFooter = () => {
+	return (
+		<Section>
+			<Container>
+				<HeadLine3>뱅카우로</HeadLine3>
+				<HeadLine3B>한우 투자하러 가볼까요?</HeadLine3B>
+				<AppLinkBox />
+			</Container>
+		</Section>
+	);
+};
 
 const Section = styled.div`
 	transition: ease all 0.5s;
 	background-color: ${(props) => props.theme.alt.backgroud1};
-	margin-top: 50px;
+	padding-top: 50px;
 	@media ${(props) => props.theme.breakpoints.sm} {
-		margin-top: 10px;
+		padding-top: 10px;
 	}
 `;
 
-const Container = styled.div`
-	width: 80%;
-	margin: 0 auto;
+const Container = styled(InnerBox)`
 	padding-left: 6%;
-	display: flex;
-	flex-direction: column;
 	@media ${(props) => props.theme.breakpoints.md} {
-		justify-content: center;
-		align-items: center;
 		padding-left: 0;
-	}
-`;
-
-const HeadLine3B = styled.h3`
-	font-weight: bold;
-	font-size: 44px;
-	color: ${(props) => props.theme.alt.text1};
-	@media ${(props) => props.theme.breakpoints.md} {
-		font-size: 36px;
-	}
-	@media ${(props) => props.theme.breakpoints.sm} {
-		width: 240px;
-		text-align: center;
 	}
 `;
 
@@ -48,16 +41,11 @@ const HeadLine3 = styled.h3`
 	}
 `;
 
-const SubFooter = () => {
-	return (
-		<Section>
-			<Container>
-				<HeadLine3>뱅카우로</HeadLine3>
-				<HeadLine3B>한우 투자하러 가볼까요?</HeadLine3B>
-				<AppLinkBox />
-			</Container>
-		</Section>
-	);
-};
+const HeadLine3B = styled(HeadLine3)`
+	font-weight: 700;
+	@media ${(props) => props.theme.breakpoints.sm} {
+		width: 240px;
+	}
+`;
 
 export default SubFooter;
