+++ src/pages/event/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useRouter } from 'next/router';
-import { Wrapper } from '../../common/typography';
+import { Wrapper } from '../../common/layouts';
 import EventDetail from '../../component/event/components/EventDetail';
 import Section1 from '../../component/event/Section1';
 import Section2 from '../../component/event/Section2';
