+++ src/pages/notice/index.tsx
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/router';
-import { Wrapper } from '../../common/typography';
+import { Wrapper } from '../../common/layouts';
 import NoticeDetail from '../../component/notice/components/NoticeDetail';
 import Section1 from '../../component/notice/Section1';
 import Section2 from '../../component/notice/Section2';
